+++ homework#1/Taewan/part1.go
